Removed handles from Cozy specification
Handles are still an experimental feature and are not quite ready for
prime-time.  This change makes the specification longer, but more
explicit about what is really happening.  It also slightly simplifies
the client code and memory management, which can now function in terms
of native types instead of Cozy-generated handle types.